# openapi/auth.yaml
openapi: 3.0.0
info:
  title: BFF API
  description: API для авторизации, регистрации и выхода
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
    description: Local development server

paths:
  /auth/register:
    post:
      summary: Регистрация нового пользователя
      description: Создает нового пользователя и устанавливает HttpOnly-куку с JWT
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          headers:
            Set-Cookie:
              description: HttpOnly-кука access_token с JWT
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с такой почтой уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Авторизация пользователя
      description: Вход в систему по почте и паролю. Устанавливает HttpOnly-куку с JWT
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthUserRequest'
      responses:
        '200':
          description: Успешная авторизация
          headers:
            Set-Cookie:
              description: HttpOnly-кука access_token с JWT
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      summary: Выход из системы
      description: Удаляет HttpOnly-куку с JWT
      tags:
        - Auth
      responses:
        '200':
          description: Успешный выход
          headers:
            Set-Cookie:
              description: Очистка куки access_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutUserResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Текущий пользователь
      description: Возвращает данные авторизованного пользователя по HttpOnly-куке
      tags:
        - Auth
      responses:
        '200':
          description: Пользователь авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Не авторизован (кука отсутствует или невалидна)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Сообщение об ошибке
          example: "Неверный email или пароль"
        statusCode:
          type: integer
          description: HTTP статус код
          example: 401

    UserResponse:
      type: object
      required: [id, email, firstName]
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "Дио"

    AuthUserRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "securePass123"

    RegisterUserRequest:
      type: object
      required: [email, password, firstName]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "securePass123"
        firstName:
          type: string
          minLength: 2
          example: "Дио"

    LogoutUserResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Успешный выход из системы"

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT в HttpOnly-куке